+++ cshelve/_azure_blob_storage.py
