+++ tests/unit/test_factory.py
@@ -9,7 +9,7 @@ from cshelve import UnknownProviderError
 from cshelve._factory import factory
 
 
-@patch("cshelve._azure.AzureBlobStorage")
+@patch("cshelve._azure_blob_storage.AzureBlobStorage")
 def test_known_backend(azure_mock):
     """
     Test factory loading the Azure backend.
